Implemented command interpreter
Created a map to store strings as keys and Actions as values. The first word introduced by the user is searched in the map and if found it uses the action value to go through a switch.

Also created a Tokenize method to split the string in words. In this method, first, I transform to lowercase the command, then I delete the spaces at the beginind and the end of the text. After that, I split the command in substrings by looking for space characters.