Implemented go command
Created a map variable in world class to with string as a key and ExitDirections enum as the value.

Created a method called "Go" in player which searchs in the entity vector for an exit in the player location with the same direction the user wants to move to and changes the player`s location.