Implemented Take Command
It checks if the item is takeable.

I checked if one of the arguments is the name of one of the items of the room. If one of them matches then I check if the item is takeable and add it to the player inventory and removes it from the room inventory