Change Go Method and exits to be hable to use keys
Now some exits will be locked and will need a key to go through them.

When used Go commando if the exit is locked and the player has the key, the exit will be unlocked and the player will go through it.

To do this I created two parameter in Exit class. A bool named locked and a Item* named key that is a pointer to the necessary Item to unlock the Exit.

I modified the Go method to search for the key in the player´s inventory and unlock the exit if found.